+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Use `ImportError` exception instead of `ModuleNotFoundError` which is not
   declared in Python 3.5 [#67][67]
 - Disallow empty JSON path [#68][68]
+- Pass the experiment to each control when requested via the `experiment`
+  parameter of the Python function [#69][69]
 
 [63]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/63
 [64]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/64
@@ -21,6 +23,7 @@
 [66]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/66
 [67]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/67
 [68]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/68
+[69]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/69
 
 ## [1.0.0rc1][] - 2018-11-30